+++ packages/components-js/src/components/basic/typography/text/text.tsx
@@ -38,7 +38,7 @@ export class Text {
   @Prop() public align?: 'left' | 'center' | 'right' = 'left';
 
   /** Basic text color variations depending on theme property. */
-  @Prop() public color?: 'brand' | 'default' | 'neutral-1' | 'neutral-2' | 'neutral-3' | 'notification-success' | 'notification-warning' | 'notification-error' | 'inherit' = 'default';
+  @Prop() public color?: 'brand' | 'default' | 'neutral-contrast-high' | 'neutral-contrast-medium' | 'neutral-contrast-low' | 'notification-success' | 'notification-warning' | 'notification-error' | 'inherit' = 'default';
 
   /** Adds an ellipsis to a single line of text if it overflows. */
   @Prop() public ellipsis?: boolean = false;